+++ src/natcap/invest/spec.py
@@ -23,7 +23,7 @@ import pint
 import pygeoprocessing
 from pygeoprocessing.utils import GDALUseExceptions
 from pydantic import AfterValidator, BaseModel, ConfigDict, \
-    field_validator, model_validator
+    field_validator, model_serializer, model_validator
 import taskgraph
 
 from natcap.invest.file_registry import FileRegistry
@@ -1066,57 +1066,23 @@ class CSVInput(FileInput):
         return [rst_line]
 
 
-class DirectoryInput(Input):
-    """A directory input, or parameter, of an invest model.
+class WorkspaceInput(Input):
+    """A workspace directory path input to an invest model.
 
-    Use this type when you need to specify a group of many file-based inputs,
-    or an unknown number of file-based inputs, by grouping them together in a
-    directory. This may also be used to describe an empty directory where model
-    outputs will be written to.
+    This is used to describe a directory where model outputs will be written.
     """
-    contents: list[Input]
-    """An iterable of `Input`s representing the contents of this directory. The
-    `key` of each input must be the file name or pattern."""
-
-    permissions: typing.Annotated[str, AfterValidator(
-        validate_permissions_string)] = ''
-    """A string that includes the lowercase characters ``r``, ``w`` and/or ``x``,
-    indicating read, write, and execute permissions (respectively) required for
-    this directory."""
-
-    must_exist: bool = True
-    """Defaults to True, indicating the directory must already exist before
-    running the model. Set to False if the directory will be created."""
-
-    type: typing.ClassVar[str] = 'directory'
-
-    rst_section: typing.ClassVar[str] = 'directory'
-
-    _contents_dict: dict[str, Input] = {}
-
-    @model_validator(mode='after')
-    def check_contents_types(self):
-        allowed_types = {
-            CSVInput, DirectoryInput, FileInput, RasterOrVectorInput,
-            SingleBandRasterInput, VectorInput}
-        for content in (self.contents or []):
-            for input_type in allowed_types:
-                if isinstance(content, input_type):
-                    break
-            else:
-                raise ValueError(
-                    f'Directory contents {content} is not an allowed type')
-        return self
-
-    def model_post_init(self, context):
-        self._contents_dict = {x.id: x for x in self.contents}
+    id: typing.ClassVar[str] = 'workspace_dir'
+    name: typing.ClassVar[str] = 'workspace directory'
+    about: str = gettext(
+        "The folder where all the model's output files will be written."
+        " If this folder does not exist, it will be created. If data"
+        " already exists in the folder, it will be overwritten.")
+    type: typing.ClassVar[str] = 'workspace'
+    rst_section: typing.ClassVar[str] = 'workspace'
 
     @property
     def display_name(self):
-        return gettext('directory')
-
-    def get_contents(self, key: str) -> Input:
-        return self._contents_dict[key]
+        return gettext('workspace directory')
 
     @timeout
     def validate(self, dirpath: str):
@@ -1128,13 +1094,6 @@ class DirectoryInput(Input):
         Returns:
             A string error message if an error was found.  ``None`` otherwise.
         """
-        if not utils._GDALPath.from_uri(dirpath).is_local:
-            return  # Don't check paths and permissions for remote paths
-
-        if self.must_exist:
-            if not os.path.exists(dirpath):
-                return validation_messages.DIR_NOT_FOUND
-
         if os.path.exists(dirpath):
             if not os.path.isdir(dirpath):
                 return validation_messages.NOT_A_DIR
@@ -1150,31 +1109,38 @@ class DirectoryInput(Input):
                     dirpath = parent
                     break
 
-        if 'r' in self.permissions:
-            try:
-                os.scandir(dirpath).close()
-            except OSError:
-                return validation_messages.NEED_PERMISSION_DIRECTORY.format(permission='read')
+        try:  # check for read permissions
+            os.scandir(dirpath).close()
+        except OSError:
+            return validation_messages.NEED_PERMISSION_DIRECTORY.format(permission='read')
 
-        # Check for x access before checking for w,
+        # Check for execute permissions before checking for write,
         # since w operations to a dir are dependent on x access
-        if 'x' in self.permissions:
-            try:
-                cwd = os.getcwd()
-                os.chdir(dirpath)
-            except OSError:
-                return validation_messages.NEED_PERMISSION_DIRECTORY.format(permission='execute')
-            finally:
-                os.chdir(cwd)
+        try:
+            cwd = os.getcwd()
+            os.chdir(dirpath)
+        except OSError:
+            return validation_messages.NEED_PERMISSION_DIRECTORY.format(permission='execute')
+        finally:
+            os.chdir(cwd)
 
-        if 'w' in self.permissions:
-            try:
-                temp_path = os.path.join(dirpath, 'temp__workspace_validation.txt')
-                with open(temp_path, 'w') as temp:
-                    temp.close()
-                    os.remove(temp_path)
-            except OSError:
-                return validation_messages.NEED_PERMISSION_DIRECTORY.format(permission='write')
+        try:  # check for write permissions
+            temp_path = os.path.join(dirpath, 'temp__workspace_validation.txt')
+            with open(temp_path, 'w') as temp:
+                temp.close()
+                os.remove(temp_path)
+        except OSError:
+            return validation_messages.NEED_PERMISSION_DIRECTORY.format(permission='write')
+
+    @model_serializer(mode="wrap")
+    def serialize(self, handler):
+        """Custom serializer to include static class attributes in the dict representation"""
+        return {
+            **handler(self),
+            'id': self.id,
+            'name': self.name,
+            'about': self.about
+        }
 
 
 class NumberInput(Input):
@@ -2301,18 +2267,7 @@ class ModelSpec(ImmutableBaseModel):
 
 
 # Specs for common arg types ##################################################
-WORKSPACE = DirectoryInput(
-    id="workspace_dir",
-    name="workspace",
-    about=(
-        "The folder where all the model's output files will be written."
-        " If this folder does not exist, it will be created. If data"
-        " already exists in the folder, it will be overwritten."
-    ),
-    contents=[],
-    permissions="rwx",
-    must_exist=False,
-)
+WORKSPACE = WorkspaceInput()
 SUFFIX = ResultsSuffixInput(
     id="results_suffix",
     name=gettext("file suffix"),